Update updater, fix some errors

+++ src/main/java/com/pantherman594/gssentials/utils/Updater.java
@@ -18,6 +18,7 @@
 
 package com.pantherman594.gssentials.utils;
 
+import com.google.common.base.Preconditions;
 import com.pantherman594.gssentials.BungeeEssentials;
 import net.md_5.bungee.api.ProxyServer;
 import net.md_5.bungee.api.plugin.Plugin;
@@ -29,6 +30,7 @@ import java.net.URLConnection;
 import java.nio.file.Files;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Map;
 import java.util.logging.Level;
 
 public class Updater {
@@ -91,7 +93,7 @@ public class Updater {
         Configuration config = BungeeEssentials.getInstance().getConfig();
         Configuration messages = BungeeEssentials.getInstance().getMessages();
         int oldVersion = getVersionFromString(config.getString("configversion"));
-        final int newVersion = getVersionFromString(plugin.getDescription().getVersion());
+        int newVersion = getVersionFromString(plugin.getDescription().getVersion());
         if (oldVersion == newVersion) {
             return;
         }
@@ -111,8 +113,12 @@ public class Updater {
             File oldConf = new File(oldDir, "config_v" + oldVersion + ".yml");
             File newMess = new File(BungeeEssentials.getInstance().getDataFolder(), "messages.yml");
             File oldMess = new File(oldDir, "messages_v" + oldVersion + ".yml");
-            Files.copy(newConf.toPath(), oldConf.toPath());
-            Files.copy(newMess.toPath(), oldMess.toPath());
+            if (!oldConf.exists()) {
+                Files.copy(newConf.toPath(), oldConf.toPath());
+            }
+            if (!oldMess.exists()) {
+                Files.copy(newMess.toPath(), oldMess.toPath());
+            }
         } catch (Exception e) {
             e.printStackTrace();
         }
@@ -145,16 +151,63 @@ public class Updater {
             oldVersion = 246;
         }
         if (oldVersion == 246) {
+            config.set("configversion", "2.4.7");
+            oldVersion = 247;
+        }
+        if (oldVersion == 247) {
             List<String> enabledList = config.getStringList("enable");
             enabledList.remove("multirelog");
             enabledList.add("autoredirect");
             enabledList.add("fastrelog");
+            enabledList.add("friend");
             enabledList.add("spam-command");
             config.set("enable", enabledList);
+            config.set("friend", Arrays.asList("friend", "f"));
+            List<Map<String, String>> section = (List<Map<String, String>>) config.getList("aliases");
+            config.set("aliases", null);
+            for (Map<String, String> map : section) {
+                Preconditions.checkNotNull(map);
+                Preconditions.checkArgument(!map.isEmpty());
+                Preconditions.checkNotNull(map.get("alias"), "invalid alias");
+                Preconditions.checkNotNull(map.get("commands"), "invalid commands");
+                config.set("aliases." + map.get("alias"), map.get("commands"));
+            }
             messages.set("multilog", null);
+            messages.set("friends.header", "&2Current Friends:");
+            messages.set("friends.body", "- {{ NAME }} ({{ SERVER }})");
+            messages.set("friends.new", "&aYou are now friends with {{ NAME }}!");
+            messages.set("friends.old", "&aYou are already friends with {{ NAME }}!");
+            messages.set("friends.remove", "&cYou are no longer friends with {{ NAME }}.");
+            messages.set("friends.outrequests.header", "&2Outgoing Friend Requests:");
+            messages.set("friends.outrequests.body", "- {{ NAME }}");
+            messages.set("friends.outrequests.new", "&a{{ NAME }} has received your friend request.");
+            messages.set("friends.outrequests.old", "&cYou already requested to be friends with {{ NAME }}. Please wait for a response!");
+            messages.set("friends.outrequests.remove", "&cThe friend request to {{ NAME }} was removed.");
+            messages.set("friends.inrequests.header", "&2Incoming Friend Requests:");
+            messages.set("friends.inrequests.body", "- {{ NAME }}");
+            messages.set("friends.inrequests.new", "&a{{ NAME }} would like to be your friend. /friend <add|remove> {{ NAME }} to accept or decline the request.");
+            messages.set("friends.inrequests.remove", "&cThe friend request from {{ NAME }} was removed.");
             messages.set("errors.fastrelog", "&cPlease wait before reconnecting!");
+            List<Map<String, String>> section2 = (List<Map<String, String>>) messages.getList("announcements");
+            config.set("aliases", null);
+            for (Map<String, String> map : section2) {
+                Preconditions.checkNotNull(map);
+                Preconditions.checkArgument(!map.isEmpty());
+                Preconditions.checkNotNull(map.get("delay"), "invalid delay");
+                Preconditions.checkNotNull(map.get("interval"), "invalid interval");
+                Preconditions.checkNotNull(map.get("message"), "invalid message");
+                String name;
+                if (config.getString("announcements." + map.get("message").split(" ")[0] + ".message") != null) {
+                    name = map.get("message").split(" ")[0];
+                } else {
+                    name = map.get("message").split(" ")[0] + Math.floor(Math.random() * 10);
+                }
+                config.set("aliases." + name + ".delay", map.get("delay"));
+                config.set("aliases." + name + ".interval", map.get("interval"));
+                config.set("aliases." + name + ".message", map.get("message"));
+            }
             config.set("configversion", "2.5.0");
-            oldVersion = 250;
+            //oldVersion = 250;
         }
         BungeeEssentials.getInstance().saveMainConfig();
         BungeeEssentials.getInstance().saveMessagesConfig();
